+++ prime-number-service-ws/src/main/java/com/company/prime/service/number/ws/model/PrimeNumberResult.java
@@ -2,14 +2,20 @@ package com.company.prime.service.number.ws.model;
 
 import java.util.List;
 
+import javax.xml.bind.annotation.XmlRootElement;
+
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 
+@XmlRootElement
 @ApiModel(description = "Container for all prime number up to the initial value")
 public class PrimeNumberResult {
 
-  private final int initial;
-  private final List<Integer> primes;
+  private int initial;
+  private List<Integer> primes;
+
+  public PrimeNumberResult() {
+  }
 
   public PrimeNumberResult(int initial, List<Integer> primes) {
     this.initial = initial;
@@ -21,9 +27,17 @@ public class PrimeNumberResult {
     return initial;
   }
 
+  public void setInitial(int initial) {
+    this.initial = initial;
+  }
+
   @ApiModelProperty("List of primes number up to initial")
   public List<Integer> getPrimes() {
     return primes;
+  }
+
+  public void setPrimes(List<Integer> primes) {
+    this.primes = primes;
   }
 
 }