+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # [0.1.0] 2018-04-05
 ### Added
 * Support different prime number verification algorithms [#2](https://github.com/ddcprg/prime-number-service/issues/2).
+### Fixed
+* Spelling mistake in list of prime number verification algorithms [#7](https://github.com/ddcprg/prime-number-service/issues/7).
 
 # [0.0.1] 2018-04-05
 ### Added