+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# [0.1.1] 2018-04-05
+### Added
+* Support XML and JSON media types [#4](https://github.com/ddcprg/prime-number-service/issues/4).
+
 # [0.1.0] 2018-04-05
 ### Added
 * Support different prime number verification algorithms [#2](https://github.com/ddcprg/prime-number-service/issues/2).